requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
tzdata==2024.1
djangorestframework==3.15.0
gunicorn==22.0.0
Markdown==3.6
packaging==24.1
pytz==2024.1
sqlparse==0.5.0
